# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MIS Backend HTTP API
  version: "1.0"
  description: |
    Specification of **actual** Micronaut application paths (prefix `/api/...`, not `/api/v1`).
    Complements the generated Swagger (`/swagger-ui`, build artifacts under `META-INF/swagger`).

    ## Authentication
    - **JWT Bearer**: `Authorization: Bearer <token>` (login response includes a `token` field).
    - **Cookie**: HttpOnly `MIS_AUTH=<token>` (set on successful `/api/auth/login` and `/api/auth/first-password-change`).
    - Anonymous: only `/api/auth/login`, `/api/auth/first-password-change`, and `GET`/`POST` under `/api/public/booking/**` (see `application.properties`).

    ## REST access control (`ApiAccessControlFilter`, `/api/**`)
    After authentication, most routes resolve to **exactly one** permission code (`x-required-permission`).
    If the code is absent (`null`), being logged in is enough. If a code is present, it must be in the employee’s effective permission set (role ± overrides via `/api/access/...`).

    Prefix → code mapping (`GET` vs everything else):
    | Prefix / condition | GET | POST, PUT, PATCH, DELETE |
    |---|---|---|
    | `/api/shell/bootstrap?kind=…` | see `kind` below | — |
    | `/api/access` | `permission.read` | `permission.write` |
    | `/api/organizations` | `organization.read` | `organization.write` |
    | `/api/branches` | `branch.read` | `branch.write` |
    | `/api/employees` | `employee.read` | `employee.write` |
...

...

...

    ## Errors
...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/branches:
    get:
      tags: [Branches]
      summary: List branches
      x-required-permission: branch.read
      parameters:
        - $ref: "#/components/parameters/pageablePage"
        - $ref: "#/components/parameters/pageableSize"
      responses:
        "200": { description: Page }
    post:
      tags: [Branches]
      summary: Create branch
      x-required-permission: branch.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
  /api/branches/{id}:
    get:
      tags: [Branches]
      summary: Branch by id
      x-required-permission: branch.read
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "200": { description: OK }
    patch:
      tags: [Branches]
      summary: Update branch
      x-required-permission: branch.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
    delete:
      tags: [Branches]
      summary: Delete branch
      x-required-permission: branch.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "204": { description: Deleted }
...
  /api/attachments:
    get:
      tags: [Attachments]
      summary: List patient attachments
      x-required-permission: patients.view
      parameters:
        - name: patientId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /api/attachments/upload:
    post:
      tags: [Attachments]
      summary: Upload file (multipart)
      description: Parts `file`, `patientId`; size and MIME limits — see `AttachmentApiController`.
      x-required-permission: patients.view
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                patientId: { type: string }
      responses:
        "200": { description: Attachment metadata }
        "400": { description: Validation error }
  /api/attachments/{id}/download:
    get:
      tags: [Attachments]
      summary: Download file (proxied from storage)
      x-required-permission: patients.view
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "200": { description: Binary body, Content-Disposition }
        "404": { description: Not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `token` field in `/api/auth/login` response
    cookieAuth:
      type: apiKey
      in: cookie
      name: MIS_AUTH
      description: HttpOnly cookie (alternative to Bearer)
  parameters:
    idUuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    pageablePage:
      name: page
      in: query
      schema: { type: integer, minimum: 0 }
      description: Page number (Micronaut Pageable)
    pageableSize:
      name: size
      in: query
      schema: { type: integer, minimum: 1 }
      description: Page size
...